@@ -17,18 +17,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.12",
-    "@angular/common": "^16.2.12",
-    "@angular/compiler": "^16.2.12",
-    "@angular/core": "^16.2.12",
+    "@angular/animations": "^17.3.11",
+    "@angular/common": "^17.3.11",
+    "@angular/compiler": "^17.3.11",
+    "@angular/core": "^17.3.11",
     "@angular/fire": "../../dist/packages-dist",
-    "@angular/forms": "^16.2.12",
-    "@angular/platform-browser": "^16.2.12",
-    "@angular/platform-browser-dynamic": "^16.2.12",
-    "@angular/platform-server": "^16.2.12",
-    "@angular/router": "^16.2.12",
-    "@angular/service-worker": "^16.2.12",
-    "@nguniversal/express-engine": "^16.2.0",
+    "@angular/forms": "^17.3.11",
+    "@angular/platform-browser": "^17.3.11",
+    "@angular/platform-browser-dynamic": "^17.3.11",
+    "@angular/platform-server": "^17.3.11",
+    "@angular/router": "^17.3.11",
+    "@angular/service-worker": "^17.3.11",
+    "@angular/ssr": "^17.3.8",
     "cross-fetch": "^3.1.4",
     "express": "^4.15.2",
     "firebase-admin": "^10.0.2",
@@ -36,18 +36,18 @@
     "rxjs": "~6.6.0",
     "tslib": "^2.1.0",
     "webpack": "^5.35.0",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1602.14",
-    "@angular-devkit/build-angular": "^16.2.14",
-    "@angular/cli": "^16.2.14",
-    "@angular/compiler-cli": "^16.2.12",
-    "@nguniversal/builders": "^16.2.0",
+    "@angular-devkit/architect": "^0.1703.8",
+    "@angular-devkit/build-angular": "^17.3.8",
+    "@angular/cli": "^17.3.8",
+    "@angular/compiler-cli": "^17.3.11",
     "@types/express": "^4.17.0",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "ajv-formats": "^2.1.1",
+    "browser-sync": "^3.0.0",
     "file-loader": "^6.2.0",
     "firebase-functions": "^3.6.0",
     "firebase-functions-test": "^0.2.2",
@@ -58,11 +58,11 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^16.2.3",
+    "ng-packagr": "^17.3.0",
     "ts-mixer": "^6.0.0",
     "ts-node": "~9.1.1",
     "tslint": "~6.1.3",
-    "typescript": "~5.1.6",
+    "typescript": "~5.4.5",
     "webpack-bundle-analyzer": "^4.4.1"
   }
 }
